Update google library version to support proxy
Come across bug with customer using the
test-report features of Codefresh. When they try to utilise it with a
Google storage bucket, they get an error relating to ssl. After some
digging, I believe I know the cause of the issue. The
cf-docker-test-reporting   image relies upon an old version of a library
to do the google-bucket uploads. The bug means accessing the google api
to do the upload fails when utilising a proxy. Bug is
https://github.com/googleapis/nodejs-storage/issues/277 . The version in
use currently is v1.7.0 . The version that fixes this issue if v2.4.0 .
It sounds like all that is needed to fix this is to update this version,
and create a new docker image.